requirements.txt
click==8.1.3
colorama==0.4.5
Flask==2.2.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector-python==8.0.30
protobuf==3.20.1
Werkzeug==2.2.1
